factor out some code
